requirements.txt
crlibm==1.0.3
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.1
numpy==1.15.2
pandas==0.23.4
pkg-resources==0.0.0
problog==2.1.0.32
pyinterval==1.2.0
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
